# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shopecart E-commerce API Documentation",
        "description": "Documentation des points d'accès (endpoints) de l'API Shopecart.",
        "contact": {
            "email": "contact@shopecart.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de l'API locale"
        }
    ],
    "paths": {
...
        "/api/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Get user's cart",
                "operationId": "3a8631f1ac30c83b5ad826df8fda966f",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1,
                                                    "nullable": true
                                                },
                                                "session_id": {
                                                    "type": "string",
                                                    "example": "abc123"
                                                },
                                                "items_count": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "total": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 99.99
                                                },
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "quantity": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "unit_price": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 29.99
                                                            },
                                                            "total": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 59.98
                                                            },
                                                            "product_variant": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer",
                                                                        "example": 1
                                                                    },
                                                                    "price": {
                                                                        "type": "number",
                                                                        "format": "float",
                                                                        "example": 29.99
                                                                    },
                                                                    "stock": {
                                                                        "type": "integer",
                                                                        "example": 10
                                                                    },
                                                                    "product": {
                                                                        "properties": {
                                                                            "id": {
                                                                                "type": "integer",
                                                                                "example": 1
                                                                            },
                                                                            "name": {
                                                                                "type": "string",
                                                                                "example": "Product Name"
                                                                            },
                                                                            "slug": {
                                                                                "type": "string",
                                                                                "example": "product-name"
                                                                            }
                                                                        },
                                                                        "type": "object"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Panier récupéré avec succès"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Create or get cart",
                "operationId": "b432c7bbb8b06ea2019c3d37f23258ca",
                "responses": {
                    "200": {
                        "description": "Cart retrieved or created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Panier récupéré"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/cart/clear": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Clear the cart",
                "operationId": "c2cf94402630acad1cc6d317ec806693",
                "responses": {
                    "200": {
                        "description": "Cart cleared successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Panier vidé avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrez le jeton Bearer obtenu après la connexion.",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}